Random Strings (for unique name ideas)
This mini program generates strings of random letters and allows the user to save ideas in a list which will ultimately be printed out in a .txt file. Each new saved list will be appended to a new line in the original file, making the file safe for multiple uses.